Overlay1.py
import streamlit as st
import cv2
import numpy as np
import geopandas as gpd
from shapely.geometry import Point, Polygon
from PIL import Image
import tempfile
import os

# Function to load and display the image
def load_image(image_file):
    img = Image.open(image_file)
    return img

# Function to create a temporary file for the KML output
def create_temp_file(suffix):
    temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=suffix)
    return temp_file.name

# Function to manually select points on the image
def select_points(img):
    st.write("Select at least three reference points on the image")
    # Display image for point selection
    points = st.image(img, use_column_width=True)
    # In reality, you'd use a tool to let the user select points here.
    # For simplicity, we assume the user knows the exact points.
    st.write("You will need to manually input coordinates")
    
    # Example points, replace this with actual user inputs
    coordinates = [(100, 150), (200, 250), (300, 350)]
    
    return coordinates

# Function to georeference the image
def georeference_image(img, points, real_coords):
    # Here, you'd implement the actual georeferencing algorithm
    # This is a placeholder showing how you'd map points to real-world coordinates.
    
    # Assume the image is perfectly aligned and scale is correct for simplicity
    h, w = img.shape[:2]
    
    # Real world coordinates corresponding to the points
    # You should input this from user data for real applications
    real_coords = [(0, 0), (0, 1), (1, 0)]
    
    return real_coords

# Main Streamlit app
st.title("Jeppesen Chart Georeferencing Tool")
st.write("Upload a Jeppesen chart to overlay on Google Earth")

image_file = st.file_uploader("Upload Image", type=["png", "jpg", "jpeg"])

if image_file is not None:
    img = load_image(image_file)
    st.image(img, caption="Uploaded Chart", use_column_width=True)

    st.write("Now, select reference points on the chart and enter their real-world coordinates.")
    points = select_points(np.array(img))

    if points:
        st.write("Enter real-world coordinates corresponding to selected points.")
        
        # Example real-world coordinates
        real_coords = [(34.000, -118.000), (34.001, -118.001), (34.002, -118.002)]
        
        st.write("Georeferencing the image...")
        geo_img = georeference_image(np.array(img), points, real_coords)

        # Create a KML file for Google Earth
        kml_file = create_temp_file(".kml")
        with open(kml_file, 'w') as f:
            f.write(f"""<?xml version="1.0" encoding="UTF-8"?>
            <kml xmlns="http://www.opengis.net/kml/2.2">
                <Document>
                    <GroundOverlay>
                        <Icon>
                            <href>{image_file.name}</href>
                        </Icon>
                        <LatLonBox>
                            <north>{real_coords[1][0]}</north>
                            <south>{real_coords[0][0]}</south>
                            <east>{real_coords[2][1]}</east>
                            <west>{real_coords[0][1]}</west>
                        </LatLonBox>
                    </GroundOverlay>
                </Document>
            </kml>""")
        
        st.write("Download the georeferenced KML file for use in Google Earth.")
        st.download_button(label="Download KML", data=open(kml_file, 'rb').read(), file_name="overlay.kml")

    else:
        st.write("Please select reference points to proceed.")
