requirements.txt
streamlit>=1.10.0
opencv-python-headless>=4.5.3.56
Pillow>=8.2.0
numpy>=1.21.0
gdal>=3.3.0
shapely>=1.7.1
geopandas>=0.9.0
fiona>=1.8.20
